refactor: create class abstraction
